6-size.c

+++ 0x00-hello_world/6-size.c
@@ -6,9 +6,9 @@
 int main(void)
 {
 	printf("Size of a char: %li byte(s)\n", (long int)sizeof(char));
-	printf("Size of an int: %li byte(s)\n", sizeof(int));
-	printf("Size of a long int: %li byte(s)\n", sizeof(long int));
-	printf("Size of a long long int: %li byte(s)\n", sizeof(long long int));
-	printf("Size of a float: %li byte(s)\n", sizeof(float));
+	printf("Size of an int: %li byte(s)\n", (long int)sizeof(int));
+	printf("Size of a long int: %li byte(s)\n", (long int)sizeof(long int));
+	printf("Size of a long long int: %li byte(s)\n", (long int)sizeof(long long int));
+	printf("Size of a float: %li byte(s)\n", (long int)sizeof(float));
 	return (0);
 }
